BG/Graveyard: Reimplement Graveyard handling as part of Map for thread safety
