+++ src/game/Globals/ObjectMgr.h
@@ -34,6 +34,7 @@
 #include "Globals/ObjectAccessor.h"
 #include "Entities/ObjectGuid.h"
 #include "Globals/Conditions.h"
+#include "Globals/GraveyardManager.h"
 
 #include <map>
 #include <climits>
@@ -344,27 +345,6 @@ struct TaxiShortcutData
 
 typedef std::unordered_multimap <uint32 /*nodeid*/, TaxiShortcutData> TaxiShortcutMap;
 
-struct GraveYardData
-{
-    uint32 safeLocId;
-    Team team;
-};
-#define GRAVEYARD_AREALINK  0
-#define GRAVEYARD_MAPLINK   1
-typedef std::multimap < uint32 /*locId*/, GraveYardData > GraveYardMap;
-typedef std::pair<GraveYardMap::const_iterator, GraveYardMap::const_iterator> GraveYardMapBounds;
-
-struct WorldSafeLocsEntry
-{
-    uint32    ID;
-    uint32    map_id;
-    float     x;
-    float     y;
-    float     z;
-    float     o;
-    char*     name;
-};
-
 struct QuestgiverGreeting
 {
     std::string text;
@@ -568,13 +548,9 @@ class ObjectMgr
         QuestgiverGreeting const* GetQuestgiverGreetingData(uint32 entry, uint32 type) const;
         TrainerGreeting const* GetTrainerGreetingData(uint32 entry) const;
 
-        WorldSafeLocsEntry const* GetClosestGraveYard(float x, float y, float z, uint32 mapId, Team team) const;
-        bool AddGraveYardLink(uint32 id, uint32 locId, uint32 linkKind, Team team, bool inDB = true);
-        void SetGraveYardLinkTeam(uint32 id, uint32 linkKey, Team team);
-        void LoadGraveyardZones();
-        GraveYardData const* FindGraveYardData(uint32 id, uint32 zoneId) const;
         void LoadWorldSafeLocs() const;
-        static uint32 GraveyardLinkKey(uint32 locId, uint32 linkKind);
+        void LoadGraveyardZones();
+        GraveYardMap& GetGraveYardMap() { return m_graveYardMap; }
 
         AreaTrigger const* GetAreaTrigger(uint32 trigger) const
         {
@@ -1248,8 +1224,6 @@ class ObjectMgr
 
         TaxiShortcutMap     m_TaxiShortcutMap;
 
-        GraveYardMap        mGraveYardMap;
-
         GameTeleMap         m_GameTeleMap;
 
         ItemRequiredTargetMap m_ItemRequiredTarget;
@@ -1275,9 +1249,8 @@ class ObjectMgr
         void LoadGossipMenuItems(std::set<uint32>& gossipScriptSet);
 
         MailLevelRewardMap m_mailLevelRewardMap;
-        WorldSafeLocsEntry const* GetClosestGraveyardHelper(
-                GraveYardMapBounds bounds, float x, float y, float z,
-                uint32 mapId, Team team) const;
+
+        GraveYardMap m_graveYardMap;
 
         typedef std::map<uint32, PetLevelInfo*> PetLevelInfoMap;
         // PetLevelInfoMap[creature_id][level]
